Add MySQL connector and modify UI in App.java
The changes in this commit add the MySQL connector dependency to the pom.xml file and modify the UI in the App.java file. The UI modification includes changing the scene to load the "login" FXML file and adjusting the scene dimensions. These changes are made to implement the user authentication feature.

Fixes #789